Navbar buttons Version1.0 (#12)
* Navbar buttons Version1.0

Add buttons(Home, Dashboard, Board and Profile). When a user is not logged in, the buttons will be hidden, and only the login/signup button will be there. Once the user is logged in, the page buttons will appear. Any log-out action will lead to going back to the home page.

* Small update

Add error funciton to useLogOut.ts
Change the isLoading to true in the useSignUp and useLogin
Delete one '!' in the AuthModal in Line 45
Remove navItems in the App.tsx

* Update useLogOut.ts

Remove the overlapped function inside the performLogOut.ts