DEMO 01 (#19)
This is to merge all of our work before our interim presentation, and demo the features below in it.

----------------------------------------------------------------------------------------------------------------------
* setting up app router (Mursal) (#1)

* Added tailwindcss and material ui + icons (#2)

* adding supabase client in the app (Mursal) (#3)

* Update Profile.tsx (#4)

* Update Profile.tsx

Add skeleton Profile UI, add a profile image, edit button, and information blocks(blocks have titles and can be editable, the data cannot be saved)

* Profile -> Update with MUI material

Add new background color as the Figma is, add new MUI TextFields and replace the old ones.

* Profile Update Edit Button

Add a new pop-up window to the edit button which users could use to upload/replace pics

* Profile Adjustments

Resize the components

* Profile.tsx Update

Remove the console.log in Function handleUploadClick().

* Setting up home page + auth foundations (Mursal) (#5)

* adding basic home page component

* changing width of app

* adding basic auth setup with hooks and modals

* fixing bugs + lint issues

* Feature/navbar + Dashboard initiated (#6)

* small updates (Imeth)

* initial api routes (Imeth)

* navbar + dashboard initiated (Imeth)

* dashboard layout (Imeth)

* scrollable projects list (Imeth)

* repeating items common component (Imeth)

* cleanup

* cleanups, del func & alert modal (Imeth)

* deleting server folder (Imeth)

* prettier single quotes (Imeth)

---------

Co-authored-by: Imeth Don <ikah0003@student.monash.edu>
Co-authored-by: Mursal Ashraf <89443356+mursal-ashraf@users.noreply.github.com>

* initial board page commit (#7)

* added initial skeleton for board page

* Merge branch 'develop' into board-page

* page styling fixes (Imeth)

* cleaned up types

---------

Co-authored-by: Imeth Don <ikah0003@student.monash.edu>

* removing server references from package json

* fixing minor bug with auth error (Mursal)

* prettier/eslint setup with Airbnb

* Feature/check is logged in (#8)

* is user logged in (Imeth)

* small comment (Imeth)

* type err fixes & review comment fixes (Imeth)

---------

Co-authored-by: Imeth Don <ikah0003@student.monash.edu>

* linting error (redundant double negation)

* updating eslint json config (Mursal)

* adding linting/formatting hooks

* formatiing

* Navbar buttons Version1.0 (#12)

* Navbar buttons Version1.0

Add buttons(Home, Dashboard, Board and Profile). When a user is not logged in, the buttons will be hidden, and only the login/signup button will be there. Once the user is logged in, the page buttons will appear. Any log-out action will lead to going back to the home page.

* Small update

Add error funciton to useLogOut.ts
Change the isLoading to true in the useSignUp and useLogin
Delete one '!' in the AuthModal in Line 45
Remove navItems in the App.tsx

* Update useLogOut.ts

Remove the overlapped function inside the performLogOut.ts

* Feature/profile Picture Uploading (#10)

* Update Profile.tsx

Add skeleton Profile UI, add a profile image, edit button, and information blocks(blocks have titles and can be editable, the data cannot be saved)

* Profile -> Update with MUI material

Add new background color as the Figma is, add new MUI TextFields and replace the old ones.

* Profile Update Edit Button

Add a new pop-up window to the edit button which users could use to upload/replace pics

* Profile Adjustments

Resize the components

* Profile.tsx Update

Remove the console.log in Function handleUploadClick().

* Profile Upload Picture function

The uploaded picture will correctly be shown in the page.

* Update Profile.tsx

Add }; at the end

---------

Co-authored-by: Mursal Ashraf <89443356+mursal-ashraf@users.noreply.github.com>

* fixing minor loading bug with hooks (Mursal)

* adding schema types from supabase (#14)

* Board Page Member Bar + Edit Card Modal (#16)

* fixed card drop bug

* refactored Board components into their own files

* added members bar + cleaned up some code

* added edit card modal

* Adding separate pages for login/signup (Mursal) (#13)

* adding login/signup pages (Mursal)

* deleting auth modal (Mursal)

* handling logged in user for auth pages

* adding common component container

* replacing auth modal in navbar

* removing lint warnings (Mursal)

* file restructure marathon

* Match Navbar file structure as with rest of codebase (Mursal)

---------

Co-authored-by: KevinLu9 <88708641+KevinLu9@users.noreply.github.com>
Co-authored-by: Dongyi Wang <100513480+Eeyore6@users.noreply.github.com>
Co-authored-by: Imeth97 <39793753+Imeth97@users.noreply.github.com>
Co-authored-by: Imeth Don <ikah0003@student.monash.edu>